rss_app/requirements.txt
Flask
Bootstrap-Flask
Flask-PyMongo
Flask-SSE
python-dotenv